Draft jobmanager
