+++ apps/src/jobmanager/pkg/config/config.py
@@ -24,15 +24,13 @@ class Config:
 
         # Broker parameters
         self.BROKER_ADDRESS = os.getenv("BROKER_ADDRESS")
-        self.BROKER_TOPIC = os.getenv("BROKER_TOPIC")
-        self.BROKER_CONSUMER_GROUP = os.getenv("BROKER_CONSUMER_GROUP")
 
     def validate(self) -> bool:
         if not self.LOGGING_LEVEL:
             return False
         if self.LOGGING_LEVEL not in ["DEBUG", "INFO", "ERROR"]:
             return False
-        
+
         # Validate database parameters
         if not self.DATABASE_MASTER_ADDRESS:
             return False
@@ -52,14 +50,9 @@ class Config:
             return False
         if not self.CACHE_PASSWORD:
             return False
-        
 
         # Validate broker parameters
         if not self.BROKER_ADDRESS:
-            return False
-        if not self.BROKER_TOPIC:
-            return False
-        if not self.BROKER_CONSUMER_GROUP:
             return False
 
         return True
