+++ apps/src/jobmanager/pkg/data/jobs.py
@@ -1,5 +1,18 @@
-import json
 from dataclasses import dataclass
+
+
+@dataclass
+class OrganizationDataObject:
+    organizationName: str
+    organizationId: str
+
+    def toDict(
+        self,
+    ):
+        return {
+            "organizationId": self.organizationId,
+            "organizationName": self.organizationName,
+        }
 
 
 @dataclass
